requirements.txt
Flask==3.0.0
Werkzeug==3.0.1
Jinja2==3.1.2
MarkupSafe==2.1.3
itsdangerous==2.1.2
click==8.1.7
blinker==1.7.0
requests
python-dotenv
supabase==2.3.0
python-jose==3.3.0
slack-sdk==3.26.0
Authlib==1.2.1